requirements.txt
streamlit
pillow
python-dotenv
google-genai
numpy
sentence-transformers
cloudinary
supabase